requirements.txt
yfinance>=0.2.28
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.9.0
reportlab>=3.6.0
openpyxl>=3.0.0
pyyaml>=6.0